Port commit dc6ad37 for List; see description
Most of the changes for this commit are commented out, because we do not want to return a singleton enumerator, due to needing to support modification during enumeration. However, this exposed a great performance opportunity of returning the Enumerator struct in GetEnumerator instead of boxing it to IEnumerable<T>, so the struct has also been made public to support this.